Update to NR solver so it check Voltage convergence instead of power.  Also added in NR iteration counter and fixed output flagging.
git-svn-id: https://svn.code.sf.net/p/gridlab-d/code/trunk@693 54b63ff4-7b51-0410-8bda-9fa96797e505

+++ powerflow/solver_nr.h
@@ -49,6 +49,6 @@ typedef struct {
 } Bus_admit;
 
 
-int solver_nr(int bus_count, BUSDATA *bus, int branch_count, BRANCHDATA *branch);
+int64 solver_nr(int bus_count, BUSDATA *bus, int branch_count, BRANCHDATA *branch);
 
 #endif
